Add ability for browser to receiver updates to already discovered services. (#19)
I was looking at adding the ability to automatically detect changes to
accessories and bridges, and needed the capability to receive TXT record
updates. The existing Browser implementation just ignores them, so this
pull request adds another event that will emit when a new message is
received for an already discovered service.

This change is 100% backwards compatible with existing implementations.